Factor back-compat sig inspection into utility func.
