+++ caproto/threading/client.py
@@ -23,7 +23,6 @@ import socket
 import sys
 import threading
 import time
-import warnings
 import weakref
 
 from queue import Queue, Empty
@@ -32,7 +31,8 @@ from collections import defaultdict, deque
 import caproto as ca
 from .._constants import (MAX_ID, STALE_SEARCH_EXPIRATION,
                           SEARCH_MAX_DATAGRAM_BYTES, RESPONSIVENESS_TIMEOUT)
-from .._utils import (batch_requests, CaprotoError, ThreadsafeCounter,
+from .._utils import (adapt_old_callback_signature,
+                      batch_requests, CaprotoError, ThreadsafeCounter,
                       socket_bytes_available, CaprotoTimeoutError,
                       CaprotoTypeError, CaprotoRuntimeError, CaprotoValueError,
                       CaprotoKeyError, CaprotoNetworkError, safe_getsockname)
@@ -2179,6 +2179,7 @@ class Subscription(CallbackHandler):
         ----------
         func : callable
             Expected signature: ``func(sub, response)``.
+
             The signature ``func(response)`` is also supported for
             backward-compatibility but will issue warnings. Support will be
             removed in a future release of caproto.
@@ -2193,40 +2194,8 @@ class Subscription(CallbackHandler):
            Changed the expected signature of ``func`` from ``func(response)``
            to ``func(sub, response)``.
         """
-        # Handle func with signature func(respons) for back-compat.
-        sig = inspect.signature(func)
-        try:
-            # Does this function accept two positional arguments?
-            sig.bind(None, None)
-        except TypeError:
-            warnings.warn(
-                "The signature of a subscription callback is now expected to "
-                "be func(sub, response). The signature func(response) is "
-                "supported, but support will be removed in a future release "
-                "of caproto.")
-            raw_func = func
-            raw_func_weakref = weakref.ref(raw_func)
-
-            def func(sub, response):
-                # Avoid closing over raw_func itself here or it will never be
-                # garbage collected.
-                raw_func = raw_func_weakref()
-                if raw_func is not None:
-                    # Do nothing with sub because the user-provided func cannot
-                    # accept it.
-                    raw_func(response)
-
-            # Ensure func does not get garbage collected until raw_func does.
-            def called_when_raw_func_is_released(w):
-                # The point of this function is to hold one hard ref to func
-                # until raw_func is garbage collected.
-                func
-                # Clean up after ourselves.
-                self.__wrapper_weakrefs.remove(w)
-
-            w = weakref.ref(raw_func, called_when_raw_func_is_released)
-            # Hold a hard reference to w. Its callback removes it from this set.
-            self.__wrapper_weakrefs.add(w)
+        # Handle func with signature func(response) for back-compat.
+        func = adapt_old_callback_signature(func, self.__wrapper_weakrefs)
 
         with self.callback_lock:
             was_empty = not self.callbacks
